reorganizing catalog utilities to be shared by resources and tools
