+++ packages/mcp/src/resources/components/store/index.ts
@@ -7,7 +7,7 @@ import { normalizeLookupValue } from "../../../shared/normalize.js";
 import {
   DEFAULT_CATALOG_SOURCE,
   createCatalogLoader,
-} from "../../shared/catalog.js";
+} from "../../../shared/catalog.js";
 import {
   getComponentLookupKeys,
   toComponentRecord,
@@ -15,7 +15,7 @@ import {
 } from "./lookup.js";
 import { componentCatalogSchema } from "./schema.js";
 
-import type { CatalogSource } from "../../shared/catalog.js";
+import type { CatalogSource } from "../../../shared/catalog.js";
 import type { ComponentRecord, ComponentSummary } from "./lookup.js";
 import type { ComponentCatalog } from "./schema.js";
