+++ packages/mcp/src/resources/flight-icons/store/index.ts
@@ -7,11 +7,11 @@ import { normalizeLookupValue } from "../../../shared/normalize.js";
 import {
   DEFAULT_CATALOG_SOURCE,
   createCatalogLoader,
-} from "../../shared/catalog.js";
+} from "../../../shared/catalog.js";
 import { getIconLookupKeys, toIconRecord } from "./lookup.js";
 import { iconCatalogSchema } from "./schema.js";
 
-import type { CatalogSource } from "../../shared/catalog.js";
+import type { CatalogSource } from "../../../shared/catalog.js";
 import type { IconRecord, IconSummary } from "./lookup.js";
 import type { IconCatalog } from "./schema.js";
