+++ packages/mcp/src/resources/tokens/store/index.ts
@@ -7,7 +7,7 @@ import { normalizeLookupValue } from "../../../shared/normalize.js";
 import {
   DEFAULT_CATALOG_SOURCE,
   createCatalogLoader,
-} from "../../shared/catalog.js";
+} from "../../../shared/catalog.js";
 import {
   getTokenLookupKeys,
   normalizeTokenLookupKey,
@@ -15,7 +15,7 @@ import {
 } from "./lookup.js";
 import { tokenCatalogSchema } from "./schema.js";
 
-import type { CatalogSource } from "../../shared/catalog.js";
+import type { CatalogSource } from "../../../shared/catalog.js";
 import type { TokenRecord, TokenSummary } from "./lookup.js";
 import type { TokenCatalogRow, TokenType } from "./schema.js";
