+++ packages/mcp/src/shared/catalog.ts
