+++ packages/mcp/tests/integration/resources/flight-icons.test.ts
@@ -14,7 +14,7 @@ import {
   parseResourceJson,
 } from "../../support/resource-content.js";
 
-import type { CatalogSource } from "../../../src/resources/shared/catalog.js";
+import type { CatalogSource } from "../../../src/shared/catalog.js";
 
 const STORE_SOURCE: CatalogSource = {
   version: "5.0.0",
