+++ packages/mcp/tests/unit/shared/catalog.test.ts
@@ -5,9 +5,9 @@
 
 import { readFileSync } from "node:fs";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { createCatalogLoader } from "../../../../src/resources/shared/catalog.js";
+import { createCatalogLoader } from "../../../../src/shared/catalog.js";
 
-import type { CatalogAnchor } from "../../../../src/resources/shared/catalog.js";
+import type { CatalogAnchor } from "../../../../src/shared/catalog.js";
 
 vi.mock("node:fs", () => ({
   readFileSync: vi.fn(),
